+++ frontend/src/types/Bytes.ts
@@ -1,6 +1,11 @@
 
 
 
+
+let currencyFormatter = new Intl.NumberFormat('en-US', {currency: 'USD', style: 'currency'});
+
+let numberFormat = new Intl.NumberFormat('en-US');
+
 export class BytesPrice {
 
     
@@ -9,14 +14,24 @@ export class BytesPrice {
 
     formattedEther: string;
     formattedUSD: string;
+
+
+    marketCap: string;
+    totalSupply: string
   
   
-    constructor(priceEther: string, priceUSD: string) {
+    constructor(priceEther: string, priceUSD: string, marketCap: string, totalSupply: string) {
         this.priceEther = Number.parseFloat(priceEther);
         this.priceUSD = Number.parseFloat(priceUSD);
 
-        this.formattedEther = this.priceEther.toFixed(5);
-        this.formattedUSD = this.priceUSD.toFixed(2);
+        this.formattedEther = numberFormat.format(this.priceEther);
+        this.formattedUSD = currencyFormatter.format(this.priceUSD);
+
+        let mc = Number.parseFloat(marketCap);
+        let ts = Number.parseFloat(totalSupply);
+
+        this.marketCap = currencyFormatter.format(mc);
+        this.totalSupply = numberFormat.format(ts);
     }
   }