+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Release v0.0.2
+---
+* Added Market Cap and Total Supply cards
+* Improved text formatting to increase readability
+
 Release v0.0.1
 ---
 * Rudimentary bytes price and Wallet UI